iOS: added #791 (separate fonts list)
Sorry for the delay getting to this.  I've implemented the separate
list of fonts, listing the sample one you provided.  Once you get the
others, they can be added to the fonts folder in the resources, and
listed in the -[FontSettingsViewController viewWillAppear:] method (I
can do that if you like).