iOS: switched story view to WKWebView
This should have significant performance improvements, leveraging the
full power of the Safari Nitro Javascript engine and out-of-process
rendering. Also as requested, it now observes the progress and removes
the spinner as soon as the base HTML has loaded.  Some big changes, but
doesn’t look like anything is broken.