iOS: kinda fixed #887 (back button text color)
I can't figure out a way to dynamically change the back button title
color; setting the appearance attributes seems to be the only way to
change it, and it doesn't do it for an already-displayed view.  So I've
changed it to use a fixed color that works well enough for all four
themes.