iOS: Fix height of font settings and font list, removing extra separators/space.
Lets the menu table views do their own height calculations and makes the bottom vertical inset match the top (negative) to avoid extra white space at the bottom of the popover.

The preferredContentSize of FontSettingsViewController was previously hard-coded to account for a navigation bar which was subsequently hidden, which could sometimes result in a too-small popover if the navigation bar was already hidden; hide it explicitly instead and remove this workaround.