Fix stories intermittently not displaying; clean up story loading.
Funnel all HTML loading through -loadHTMLString:.

Cache base URL (which is always the same) in a static variable instead of computing it different ways and putting in an instance variable temporarily.

Use more intention-revealing variable names.

Only set hasStory when we’re actually loading the story itself.

Only set the font size and line spacing once, in webViewDidStartLoad, because the DOM is always set up by then in my experience.

Only bother to set font size and line spacing if we’re loading a (full) story.

Show the Web view when a full story *starts* to load, as sometimes it may not complete but you still want to see a partially loaded story.  (Fixes an intermittent issue with AnandTech; see https://getsatisfaction.com/newsblur/topics/anandtech-straight-up-doesnt-load-on-the-ios-app)