iOS: Replace “magic number” 0-3 section indices with constants.
These indices are used in appDelegate.dictFoldersArray and as UIButton tags.

Removes a number of places where a number-to-name lookup was done for no apparent reason when the items are in appDelegate.dictFoldersArray already.

Fixes issue where all stories are now at index 3, not 2, and shift-E was showing infrequent stories on iPad because the index was not updated when infrequent stories were introduced.