iOS: #1137 (full screen reading)
Added a "Tap story" preference with "Toggle full screen" and "Do nothing" options; toggling is the default.  This is only available on iPhone.  When active, a single tap in a story, not on a link, image, movie, or button, will hide or show the status and navigation bars.  On iPhones with a notch (iPhone X, XS, etc), the feed bar at the top is also hidden, so the story can scroll under the notch, using every bit of the screen.