iOS: done #1156 (story titles height)
Replaced the Preview descriptions preference with a multi-select for title only, plus short, medium, and long title & content. The text space is dynamically distributed between the two. The thumbnail is constrained in size so both the large and small variations work with all preview text sizes, with the date below if it fits.