Remove mobile check for update_counts
This check was added four years ago in .50e88b2bafd8b183b364484b57d3ed455db7b27a. The expectation was that the mobile clients would be migrated to using a combination of load_feeds and refresh_feeds. However, due to a typo, this check was only applied to Apple (iPad, iOS) clients. In the interim, we've ended up in the following situation:

- The iOS client uses refresh_feeds, as it has for 12 years since 09affc6ccf49e5520a90a623412e28a829666270
- The Android client uses load_feeds with update_counts=True, as it has been allowed to get away with for years...

Let's bring back equality for iOS clients!